fix: resolve rebase conflict in proxies.ts — keep hasBlockingProxyAssignment but drop duplicate extraction leftovers
- Removed duplicate resolveScopePoolInternal, resolveProxyForConnectionFromRegistry,
  resolveProxyForScopeFromRegistry already extracted to proxies/rotation.ts
- Removed duplicate hasBlockingProxyAssignment function body already re-exported from proxies/guards.ts
- Removed duplicate PROXY_ALIVE_PREDICATE import
- All typechecks and 45 affected tests pass
