+++ config/quality/file-size-baseline.json
@@ -250,9 +250,8 @@
     "src/lib/db/core.ts": 1825,
     "src/lib/db/migrationRunner.ts": 1125,
     "src/lib/db/models.ts": 1259,
-    "_rebaseline_2026_07_15_perf_api_pagination": "PR #7046 (perf/api-pagination) own growth: proxies.ts 1177->1208 (+31) — backward-compatible pagination refactor: listProxies returns plain array when called without limit/offset (preserving existing callers), {items,total} when pagination requested. Branch-local baseline bump for CI fast-gates.",
     "src/lib/db/providers.ts": 1107,
-    "src/lib/db/proxies.ts": 1208,
+    "src/lib/db/proxies.ts": 1177,
     "src/lib/db/settings.ts": 1155,
     "src/lib/db/usageAnalytics.ts": 925,
     "src/lib/evals/evalRunner.ts": 961,
